Fix parsing of category labels

+++ src/common/timeseries/category-selection/category-selector.component.ts
@@ -191,9 +191,9 @@ export class CategorySelectorComponent implements OnInit {
    */
   private prepareOrderedCategoryGroup(result: ExtendedTimeseries, asc: Boolean): TimeSeriesGroup {
     const length = this.categoryList.length;
-    const resValue = parseInt(result.parameters.category.label.slice(0, -1));
+    const resValue = parseFloat(result.parameters.category.label.slice(0, -1));
       for (var _i = 0; _i < length; _i++) {
-        let actValue = parseInt(this.categoryList[_i].label.slice(0, -1));
+        let actValue = parseFloat(this.categoryList[_i].label.slice(0, -1));
         if(resValue === actValue){
           let group = this.categoryList[_i];
           group.timeseries.push(result);
@@ -219,12 +219,12 @@ export class CategorySelectorComponent implements OnInit {
    */
   private prepareOrderedPhenomenonGroup(result: ExtendedTimeseries, asc: Boolean): TimeSeriesGroup {
     const length = this.phenomenonList.length;
-    const resValue = parseInt(result.parameters.category.label.slice(0, -1));
+    const resValue = parseFloat(result.parameters.category.label.slice(0, -1));
       for (var _i = 0; _i < length; _i++) {
         if(result.parameters.phenomenon.label === this.phenomenonList[_i].label){
           const timeseriesLength = this.phenomenonList[_i].timeseries.length;
           for(var _j = 0; _j < timeseriesLength; _j++){
-            let actValue = parseInt(this.phenomenonList[_i].timeseries[_j].parameters.category.label.slice(0, -1));
+            let actValue = parseFloat(this.phenomenonList[_i].timeseries[_j].parameters.category.label.slice(0, -1));
             if(resValue === actValue){
               let group = this.phenomenonList[_i];
               this.phenomenonList[_i].timeseries.push(result);
